Restructure broker func. Add return after error
