+++ pkg/broker/broker.go
@@ -2,26 +2,7 @@ package broker
 
 import (
 	"github.com/codex-team/hawk.collector/cmd"
-	"github.com/streadway/amqp"
 )
-
-// Connection contains basic settings for AMQP connection
-type Connection struct {
-	// Exchange name
-	Exchange string
-
-	// Queue name
-	Queue string
-
-	// If message must be routable (http://www.rabbitmq.com/faq.html#mandatory-flat-routing)
-	Mandatory bool
-
-	// If message must be delivered only if there is a consumer (http://www.rabbitmq.com/faq.html#immediate-flat-routing)
-	Immediate bool
-
-	// Channel link
-	channel *amqp.Channel
-}
 
 // Message represents message payload sent to the Queue and AMQP route
 type Message struct {
